Refactor server to be fully code generated
