+++ rust-runtime/aws-smithy-http-server-python/src/lib.rs
@@ -13,7 +13,6 @@
 
 mod error;
 mod logging;
-mod server;
 mod socket;
 mod state;
 pub mod types;
@@ -22,8 +21,6 @@ pub mod types;
 pub use error::Error;
 #[doc(inline)]
 pub use logging::{setup, LogLevel};
-#[doc(inline)]
-pub use server::PyServer;
 #[doc(inline)]
 pub use socket::SharedSocket;
 #[doc(inline)]
